clients/web: remove IssueReference and IssueDependency logic
