+++ clients/apps/web/src/app/[organization]/[repo]/issues/[number]/(create)/page.tsx
@@ -1,12 +1,6 @@
 import { getServerSideAPI } from '@/utils/api'
 import { redirectToCanonicalDomain } from '@/utils/nav'
-import {
-  Issue,
-  Pledger,
-  PullRequest,
-  ResponseError,
-  RewardsSummary,
-} from '@polar-sh/sdk'
+import { Issue, Pledger, ResponseError, RewardsSummary } from '@polar-sh/sdk'
 import { Metadata, ResolvingMetadata } from 'next'
 import { headers } from 'next/headers'
 import { notFound } from 'next/navigation'
@@ -81,7 +75,6 @@ export default async function Page({
   let issueHTMLBody: string | undefined
   let pledgers: Pledger[] = []
   let rewards: RewardsSummary | undefined
-  let pulls: PullRequest[] = []
 
   const api = getServerSideAPI()
 
@@ -93,20 +86,17 @@ export default async function Page({
       cacheConfig,
     )
 
-    const [bodyResponse, pledgeSummary, rewardsSummary, pullRequests] =
-      await Promise.all([
-        api.issues.getBody({ id: issue.id }, { next: { revalidate: 60 } }), // Cache for 60s
-        api.pledges.summary({ issueId: issue.id }, cacheConfig),
-        api.rewards.summary({ issueId: issue.id }, cacheConfig),
-        api.pullRequests.search({ referencesIssueId: issue.id }, cacheConfig),
-      ])
+    const [bodyResponse, pledgeSummary, rewardsSummary] = await Promise.all([
+      api.issues.getBody({ id: issue.id }, { next: { revalidate: 60 } }), // Cache for 60s
+      api.pledges.summary({ issueId: issue.id }, cacheConfig),
+      api.rewards.summary({ issueId: issue.id }, cacheConfig),
+    ])
 
     issueHTMLBody = bodyResponse
     pledgers = pledgeSummary.pledges
       .map(({ pledger }) => pledger)
       .filter((p): p is Pledger => !!p)
     rewards = rewardsSummary
-    pulls = pullRequests.items || []
   } catch (e) {
     if (e instanceof ResponseError && e.response.status === 404) {
       notFound()
@@ -131,7 +121,6 @@ export default async function Page({
       pledgers={pledgers}
       rewards={rewards}
       gotoURL={undefined}
-      pullRequests={pulls}
     />
   )
 }
