+++ clients/apps/web/src/components/Pledge/Status.tsx
@@ -81,7 +81,6 @@ export const Status = (props: {
           <IssueListItem
             issue={pledge.issue}
             pledges={[pledge]}
-            references={[]}
             checkJustPledged={false}
             canAddRemovePolarLabel={false}
             showPledgeAction={true}
