+++ clients/packages/polarkit/src/components/Issue/ListItemDecoration.tsx
@@ -1,36 +1,28 @@
 import {
   Funding,
   Issue,
-  IssueReferenceRead,
   Pledge,
   PledgeState,
   PledgesTypeSummaries,
   Reward,
 } from '@polar-sh/sdk'
 import { getCentsInDollarString } from 'polarkit/money'
-import { twMerge } from 'tailwind-merge'
 import IssuePledge from './IssuePledge'
-import IssueReference from './IssueReference'
 import IssueRewards from './IssueRewards'
 
 // When rendering in the Chrome Extension, the iframe needs to know it's expected height in pixels
 export const getExpectedHeight = ({
   pledges,
-  references,
 }: {
   pledges: Pledge[]
-  references: IssueReferenceRead[]
 }): number => {
   const pledgeHeight = pledges.length > 0 ? 28 : 0
-  const referenceHeight = 28 * references.length
-  const inner = Math.max(pledgeHeight, referenceHeight)
-  return inner + 24
+  return pledgeHeight + 24
 }
 
 const IssueListItemDecoration = ({
   pledges,
   pledgesSummary,
-  references,
   showDisputeAction,
   onDispute,
   onConfirmPledges,
@@ -42,7 +34,6 @@ const IssueListItemDecoration = ({
 }: {
   pledges: Array<Pledge>
   pledgesSummary?: PledgesTypeSummaries
-  references: IssueReferenceRead[]
   showDisputeAction: boolean
   onDispute: (pledge: Pledge) => void
   onConfirmPledges: () => void
@@ -117,8 +108,6 @@ const IssueListItemDecoration = ({
     return pledge.amount.amount
   }
 
-  const haveReferences = references && references.length > 0
-
   const pledgesSummaryOrDefault = pledgesSummary ?? {
     pay_directly: { total: { currency: 'USD', amount: 0 }, pledgers: [] },
     pay_on_completion: { total: { currency: 'USD', amount: 0 }, pledgers: [] },
@@ -138,25 +127,6 @@ const IssueListItemDecoration = ({
             confirmPledgeIsLoading={confirmPledgeIsLoading}
             funding={funding}
           />
-        )}
-        {haveReferences && (
-          <div
-            className={twMerge(
-              'dark:bg-polar-900 space-y-2 bg-gray-50 px-6 py-2',
-            )}
-          >
-            {haveReferences &&
-              references.map((r: IssueReferenceRead) => {
-                return (
-                  <IssueReference
-                    orgName={issue.repository.organization.name}
-                    repoName={issue.repository.name}
-                    reference={r}
-                    key={r.id}
-                  />
-                )
-              })}
-          </div>
         )}
         {rewards && rewards?.length > 0 && <IssueRewards rewards={rewards} />}
       </div>
